pyproject.toml
[tool.poetry]
name = "hw05_final"
version = "0.1.0"
description = ""
authors = ["SaPer663 <saper663ru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "2.2.16"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "4.7.2"
PyJWT = "2.1.0"
djoser = "^2.1.0"
drf-yasg = "^1.20.0"
Pillow = "8.3.1"
requests = "2.26.0"
six = "1.16.0"
sorl-thumbnail = "12.7.0"
pytils = "^0.3"
psycopg2-binary = "2.8.6"
gunicorn = "^20.0"
python-decouple = "^3.5"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.10.1"
ipython = "^7.29.0"
django-debug-toolbar = "3.2.4"
coverage = "^6.2"
pytest = "6.2.4"
pytest-django = "4.4.0"
pytest-pythonpath = "0.7.3"
mixer = "7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
